refactor: break down logic into service files
